backend/requirements.txt
fastapi[all]
uvicorn[standard]
SQLAlchemy
psycopg2
alembic
python-dotenv
pytest
websockets
python-jose[cryptography]
passlib[bcrypt]
sse-starlette